+++ src/main/java/com/a3/catalogofilme/repository/FilmeRepository.java
@@ -3,5 +3,5 @@ package com.a3.catalogofilme.repository;
 import com.a3.catalogofilme.model.Filme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmeRepository extends JpaRepository<Filme, Long> {
+public interface FilmeRepository extends JpaRepository<Filme, Integer> {
 }
